Changed background tasks to also run while the app is in the foreground
